requirements.txt
fastapi
uvicorn
pandas
openpyxl
google-genai
google-api-core
pydantic
rapidfuzz
python-dotenv
tenacity
PyYAML
markdown-it-py
sqlalchemy>=2.0
psycopg2-binary
alembic
asyncpg
pgvector
numpy
pytest